feat: eliminate builder
